Removed unused classes and traits
Removed Article, PublishDateExtractor and VoicesContentExtractor as they were not used.

Also, removed the traits and the configuration so now Gander uses always the SandardContentExtractor
(now called ContentExtractor) or StandardDocumentCleaner (now just DocumentCleaner).
If someday we want to support another ContentExtractor, we could easily do it by passing functions
instead of the more Java way of creating interfaces (traits) and subclassing.