Added integration tests for links
Gander was collecting the links of the main content but
we were not covering it by tests.

Improved README.md

Also, removed folha test as it fails because of the
charset. I've tried all the charsets (including the one
specified in the html) but it keeps failing. As it does not
add many value I've decided to kill it.